Added a whole bunch of functions that are necessary to automatically find the spots, spot centers, spot fits, and windowing function parameters from the FT of a model.
It is expected that the FT was created using 512 pixels, but the IFTs
will be done using 256 pixels.

Functions to create the parameter file are also included.

"Fixing" the fitting parameters for incorrectly fitted spots is in
version 0.1. You may need to fix each "wrong" spot manually. Some
functions to figure out how to do this are included. However, I have
created one function to attempt to "fix" the fit automatically. Use with
caution. There is a reason it is called "try_automagic_spot_fix", I have
not tested it well at all, and I don't expect it to work fantastically.
It might be better than the poor fit you are trying to fix though.