enable many basis atoms
Modify BasisParameters and CreateGridWithBasis to allow any number of
basis atoms, not just one.  The vectors from a lattice point to each
basis atom are stored in the columns of c_params.

Also fixed a fencepost error in the number of atoms in
CreateGridWithBasis and added EstimateVector to pre-calculate the inputs
to BasisParameters from the cursors on a grid.