CreateGridWithBasis updates
bugfix in the number of atoms for CreateGridWithBasis.

changed centering behavior for created grid.  Previous behavior was to
match the grid to the atom closest to (0, 0).  Now the grid is shifted
so it's center is the same as the center of the x0, y0 atom list, then
shifted again to match up exactly the closest positions in the two
lists.  This behavior DOES NOT WORK for complicated atom bases.