added "no noise" option to GaussianFit
Added "no noise" optional parameter to GaussianFit and GausianFitStack.
Noise is calculated as the sqrt of the input image, assuming that the
input is in detected electrons.  If that is not true, using this
definition for the noise leads to spurious results.  If some pixels in
the input image are negative, the fit will fail because some of the
noise pixels are NaN.

Not all combination of input parameters and optional parameters are
coded.  For example, neither "no noise" nor "fix xy" works with a
non-zero "wiggle" parameter for the fit.  Nor does "no noise" without
"fix xy" work.