first commit of ECM routines, LoadTiff & relaxation
LoadTiff loads an entire directory of tif files into a 3D matrix.  This
version is much faster than the previous version.  The previous version
is not in the repo.
relaxation contains functions for ECM analysis of a time series of
diffraction patterns, including calculating the time autocorrelation
function g2 and fitting to the KWW equation.